refactor: Separate middlewares
